// enums.ts
export enum AppointmentStatus {
  PENDING = "Pending",
  CONFIRMED = "Confirmed",
  CANCELLED = "Cancelled",
}

export enum Gender {
  MALE = "Male",
  FEMALE = "Female",
  OTHER = "Other",
}

export enum Language {
  ENGLISH = "English",
  FRENCH = "French",
  ARABIC = "Arabic",
}

export enum CustomerStatus {
  ACTIVE = "Active",
  INACTIVE = "Inactive",
}

export enum SESSION_STEPS {
  SESSION_FORMAT = "SESSION_FORMAT",
  PREPARE = "prepare",
  RECORDING = "recording",
  FIND_PATIENT = "FIND_PATIENT",
  PROCESSING = "processing",
  END = "end",
}